switch from bun to node

+++ src/index.ts
@@ -1,12 +1,10 @@
 import { TelegramClient, Api } from 'telegram';
-import { StringSession } from 'telegram/sessions';
-import input from 'input'; // npm i input
+import input from 'input';
 import * as fs from 'fs';
-import { EntityLike } from 'telegram/define';
-import { NewMessage, NewMessageEvent, Raw } from 'telegram/events';
+import { NewMessageEvent } from 'telegram/events';
 import { Dialog } from 'telegram/tl/custom/dialog';
-import { dialogs } from 'telegram/client';
-import natural from 'natural';
+import * as natural from 'natural';
+import { StringSession } from 'telegram/sessions/index.js';
 
 const APP_ID = parseInt(process.env.APP_ID || '');
 const API_HASH = process.env.API_HASH;
@@ -17,7 +15,7 @@ const data = { stopWords: [], banCounter: 0, restart: 0, restrictedLetters: [] }
 
 const loadData = () => {
     try {
-        const savedData = JSON.parse(fs.readFileSync('persist', 'utf-8'));
+        const savedData = JSON.parse(fs.readFileSync('./.persist', 'utf-8'));
         Object.keys(data).forEach(key => {
             if (key in savedData) data[key] = savedData[key];
         });
@@ -25,14 +23,14 @@ const loadData = () => {
 };
 
 const saveData = () => {
-    fs.writeFileSync('persist', JSON.stringify(data, null, 2));
+    fs.writeFileSync('./.persist', JSON.stringify(data, null, 2));
 };
 
 const initializeClient = async (): Promise<TelegramClient> => {
     if (!APP_ID) throw new Error('!process.env.APP_ID');
     if (!API_HASH) throw new Error('!process.env.API_HASH');
 
-    const stringSession = fs.readFileSync('./session', {
+    const stringSession = fs.readFileSync('./.session', {
         encoding: 'utf-8',
     });
 
@@ -47,20 +45,14 @@ const initializeClient = async (): Promise<TelegramClient> => {
         onError: err => console.log(err),
     });
 
-    const stringSessionNew: string = client.session.save();
-    fs.writeFileSync('./session', stringSessionNew, 'utf-8');
+    const stringSessionNew: string = client.session.save() as unknown as string;
+
+    fs.writeFileSync('./.session', stringSessionNew, 'utf-8');
 
     loadData();
     data.restart++;
     console.log(data);
     setInterval(saveData, 1000);
-
-    // client.invoke(
-    //     new Api.messages.DeleteMessages({
-    //         id: [event.message.id],
-
-    //     }),
-    // );
 
     return client;
 };
@@ -112,19 +104,27 @@ initializeClient().then(async client => {
                     }),
                 );
 
-                // if (result.count > 10) return;
+                //@ts-ignore
+                const count: string | number = result.count || '---';
+
+                if (typeof count === 'number' && count > 10) return;
 
                 data.banCounter++;
-                const message = `Сообщение удалено за ${reason}. Кол-во сообщений от пользователя ${result.count}`;
+
+                const message = `Сообщение удалено за ${reason}. Кол-во сообщений от пользователя ${count}`;
 
                 client.deleteMessages(dialog.id, [event.message.id], {
                     revoke: true,
                 });
 
-                client.sendMessage(dialog.id!, {
+                const message2 = await client.sendMessage(dialog.id!, {
                     message,
                     silent: true,
                 });
+
+                setTimeout(() => {
+                    message2.delete();
+                }, 60 * 1000);
             }
         } else if (event instanceof Api.UpdateUserStatus) {
             console.log(`UpdateUserStatus`);
@@ -133,43 +133,12 @@ initializeClient().then(async client => {
                 if (typeof value === 'function' && event instanceof value)
                     console.log(`instanceof ${key}`);
             });
-            // console.log(`unknown`, event);
         }
-        // fs.writeFileSync('./log', '\n' + JSON.stringify(event, null, 2), {
-        //     flag: 'a+',
-        // });
-        // Use the args from event.message.patternMatch..
-        // await event.message.reply({ text: "Thanks, registered!" });
     });
 
     const dialogs = await client.getDialogs();
 
     dialog = dialogs.find(x => x.name === 'МойСклад API: сообщество разработчиков')!;
-
-    // const result = await client.invoke<Api.messages.Search>(
-    //     new Api.messages.Search({
-    //         peer: dialog.id,
-    //         fromId: 'another_star',
-    //         q: '',
-    //         filter: new Api.InputMessagesFilterEmpty(),
-    //     }),
-    // );
-
-    // console.log(result);
-
-    // const result = await client.invoke(
-    //     new Api.messages.GetHistory({
-    //         peer: dialog.id,
-    //         //   offsetId: 43,
-    //         //   offsetDate: 43,
-    //         addOffset: 24900,
-    //         limit: 1,
-    //         //   maxId: 0,
-    //         //   minId: 0,
-    //     }),
-    // );
-
-    // console.log(result);
 });
 
 const sendMessage = async (message: string) => {
@@ -181,27 +150,3 @@ const sendMessage = async (message: string) => {
         }),
     );
 };
-
-// const stringSessionOld = fs.readFileSync('./session', {
-//     encoding: 'utf-8',
-// });
-
-// (async () => {
-
-//     await client.start({
-//         phoneNumber: async () => await input.text('Please enter your number: '),
-//         password: async () => await input.text('Please enter your password: '),
-//         phoneCode: async () => await input.text('Please enter the code you received: '),
-//         onError: err => console.log(err),
-//     });
-
-//     console.log('You should now be connected.');
-
-//     const stringSessionNew: string = client.session.save();
-
-//     console.log({ stringSessionNew });
-
-//     fs.writeFileSync('./session', stringSessionNew, 'utf-8');
-
-//     await client.sendMessage('me', { message: 'Hello!' });
-// })();
